refactor flashCardsAPI.ts, fractured in separate services
