+++ services/decks/decks.service.ts
@@ -1,0 +1,102 @@
+import { flashCardsAPI } from '../flashCardsAPI'
+import {
+  CreateDeckArgs,
+  Deck,
+  DecksListResponse,
+  DeleteDeckArgs,
+  GetDecksArgs,
+  UpdateDeckArgs,
+} from './deck.types'
+
+export const decksService = flashCardsAPI.injectEndpoints({
+  endpoints: builder => {
+    return {
+      createDeck: builder.mutation<Deck, CreateDeckArgs>({
+        invalidatesTags: ['Deck'],
+        query: ({ cover, isPrivate, name }) => {
+          const formData = new FormData()
+
+          if (name) {
+            formData.append('name', name)
+          }
+          if (isPrivate) {
+            formData.append('isPrivate', isPrivate.toString())
+          }
+          if (cover) {
+            formData.append('cover', cover)
+          } else if (cover === null) {
+            formData.append('cover', '')
+          }
+
+          return {
+            body: formData,
+            method: 'POST',
+            url: 'v1/decks',
+          }
+        },
+      }),
+      deleteDeck: builder.mutation<void, DeleteDeckArgs>({
+        invalidatesTags: ['Deck'],
+        query: ({ id, ...args }) => ({
+          body: args,
+          method: 'DELETE',
+          url: `v1/decks/${id}`,
+        }),
+      }),
+      getDeckById: builder.query<Deck, { id: string }>({
+        providesTags: ['Deck'],
+        query: ({ id }) => ({
+          method: 'GET',
+          url: `v1/decks/${id}`,
+        }),
+      }),
+      getDecks: builder.query<DecksListResponse, GetDecksArgs | void>({
+        providesTags: ['Deck'],
+        query: args => ({
+          method: 'GET',
+          params: {
+            ...(args ?? {}),
+            authorId: args?.authorId || undefined,
+            currentPage: args?.currentPage || undefined,
+            itemsPerPage: args?.itemsPerPage || undefined,
+            name: args?.name || undefined,
+            orderBy: args?.orderBy || undefined,
+          },
+          url: 'v2/decks',
+        }),
+      }),
+      updateDeck: builder.mutation<Deck, UpdateDeckArgs>({
+        invalidatesTags: ['Deck'],
+        query: ({ cover, id, isPrivate, name }) => {
+          const formData = new FormData()
+
+          if (name) {
+            formData.append('name', name)
+          }
+          if (isPrivate) {
+            formData.append('isPrivate', isPrivate.toString())
+          }
+          if (cover) {
+            formData.append('cover', cover)
+          } else if (cover === null) {
+            formData.append('cover', '')
+          }
+
+          return {
+            body: formData,
+            method: 'PATCH',
+            url: `v1/decks/${id}`,
+          }
+        },
+      }),
+    }
+  },
+})
+
+export const {
+  useCreateDeckMutation,
+  useDeleteDeckMutation,
+  useGetDeckByIdQuery,
+  useGetDecksQuery,
+  useUpdateDeckMutation,
+} = decksService
