+++ services/flashCardsAPI.ts
@@ -1,15 +1,7 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react'
 
-import {
-  CardsListResponse,
-  CreateDeckArgs,
-  Deck,
-  DecksListResponse,
-  DeleteDeckArgs,
-  GetCardsArgs,
-  GetDecksArgs,
-  UpdateDeckArgs,
-} from './decks/deck.types'
+import { CardsListResponse, GetCardsArgs } from './decks/deck.types'
+
 export const flashCardsAPI = createApi({
   baseQuery: fetchBaseQuery({
     baseUrl: 'https://api.flashcards.andrii.es',
@@ -20,38 +12,6 @@ export const flashCardsAPI = createApi({
   }),
   endpoints: builder => {
     return {
-      createDeck: builder.mutation<Deck, CreateDeckArgs>({
-        invalidatesTags: ['Deck'],
-        query: ({ cover, isPrivate, name }) => {
-          const formData = new FormData()
-
-          if (name) {
-            formData.append('name', name)
-          }
-          if (isPrivate) {
-            formData.append('isPrivate', isPrivate.toString())
-          }
-          if (cover) {
-            formData.append('cover', cover)
-          } else if (cover === null) {
-            formData.append('cover', '')
-          }
-
-          return {
-            body: formData,
-            method: 'POST',
-            url: 'v1/decks',
-          }
-        },
-      }),
-      deleteDeck: builder.mutation<void, DeleteDeckArgs>({
-        invalidatesTags: ['Deck'],
-        query: ({ id, ...args }) => ({
-          body: args,
-          method: 'DELETE',
-          url: `v1/decks/${id}`,
-        }),
-      }),
       getCards: builder.query<CardsListResponse, { args: GetCardsArgs; id: string }>({
         providesTags: ['Card'],
         query: ({ args, id }) => ({
@@ -67,62 +27,9 @@ export const flashCardsAPI = createApi({
           url: `v1/decks/${id}/cards`,
         }),
       }),
-      getDeckById: builder.query<Deck, { id: string }>({
-        providesTags: ['Deck'],
-        query: ({ id }) => ({
-          method: 'GET',
-          url: `v1/decks/${id}`,
-        }),
-      }),
-      getDecks: builder.query<DecksListResponse, GetDecksArgs | void>({
-        providesTags: ['Deck'],
-        query: args => ({
-          method: 'GET',
-          params: {
-            ...(args ?? {}),
-            authorId: args?.authorId || undefined,
-            currentPage: args?.currentPage || undefined,
-            itemsPerPage: args?.itemsPerPage || undefined,
-            name: args?.name || undefined,
-            orderBy: args?.orderBy || undefined,
-          },
-          url: 'v2/decks',
-        }),
-      }),
-      updateDeck: builder.mutation<Deck, UpdateDeckArgs>({
-        invalidatesTags: ['Deck'],
-        query: ({ cover, id, isPrivate, name }) => {
-          const formData = new FormData()
-
-          if (name) {
-            formData.append('name', name)
-          }
-          if (isPrivate) {
-            formData.append('isPrivate', isPrivate.toString())
-          }
-          if (cover) {
-            formData.append('cover', cover)
-          } else if (cover === null) {
-            formData.append('cover', '')
-          }
-
-          return {
-            body: formData,
-            method: 'PATCH',
-            url: `v1/decks/${id}`,
-          }
-        },
-      }),
     }
   },
   tagTypes: ['Deck', 'Card'],
 })
 
-export const {
-  useCreateDeckMutation,
-  useDeleteDeckMutation,
-  useGetCardsQuery,
-  useGetDeckByIdQuery,
-  useGetDecksQuery,
-  useUpdateDeckMutation,
-} = flashCardsAPI
+export const { useGetCardsQuery } = flashCardsAPI
