+++ src/pagesMinin/deckCardsPage.module.scss
