+++ src/pagesMinin/utls/variablesMinin.ts
